python/requirements.txt
pyserial>=3.5
numpy>=1.21.0
matplotlib>=3.3.0
tkinter-page>=0.0.5
PyQt5>=5.15